feat: listbox in table
